feat: Creating mock-up pages for Login/Register pages
